+++ contracts/vara-man/game/src/contract.rs
@@ -1,3 +1,4 @@
+use super::ft_messages;
 use gstd::{exec, msg, prelude::*, ActorId};
 use hashbrown::HashMap;
 use vara_man_io::{
@@ -11,6 +12,21 @@ struct VaraMan {
     pub players: HashMap<ActorId, Player>,
     pub status: Status,
     pub config: Config,
+    pub transaction_id: u64,
+    pub transactions: HashMap<u64, Option<VaraManAction>>,
+}
+
+impl VaraMan {
+    pub fn lazy_get_transaction_id(&mut self, transaction_id: Option<u64>) -> u64 {
+        match transaction_id {
+            Some(transaction_id) => transaction_id,
+            None => {
+                let transaction_id = self.transaction_id;
+                self.transaction_id = self.transaction_id.wrapping_add(1);
+                transaction_id
+            }
+        }
+    }
 }
 
 impl From<&VaraMan> for VaraManState {
@@ -35,18 +51,36 @@ async fn main() {
     let action: VaraManAction = msg::load().expect("Unexpected invalid action payload.");
     let vara_man: &mut VaraMan = unsafe { VARA_MAN.get_or_insert(VaraMan::default()) };
 
-    let result = process_handle(action, vara_man).await;
+    if let VaraManAction::ClaimReward {
+        game_id: _,
+        silver_coins: _,
+        gold_coins: _,
+    } = &action
+    {
+        vara_man
+            .transactions
+            .insert(vara_man.transaction_id, Some(action.clone()));
+    }
 
+    let result = process_handle(None, action, vara_man).await;
     msg::reply(result, 0).expect("Unexpected invalid reply result.");
 }
 
-async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraManEvent {
+async fn process_handle(
+    transaction_id: Option<u64>,
+    action: VaraManAction,
+    vara_man: &mut VaraMan,
+) -> VaraManEvent {
     match action {
         VaraManAction::RegisterPlayer { name } => {
             let actor_id = msg::source();
 
             if vara_man.status == Status::Paused {
                 return VaraManEvent::Error("Incorrect whole game status.".to_owned());
+            }
+
+            if name.is_empty() {
+                return VaraManEvent::Error("Username is empty.".to_owned());
             }
 
             if vara_man
@@ -84,8 +118,10 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
 
             player.retries += 1;
 
-            vara_man.games.push(GameInstance::new(
+            vara_man.games.push(GameInstance::new_with_coins(
                 level,
+                vara_man.config.gold_coins,
+                vara_man.config.silver_coins,
                 player_address,
                 exec::block_timestamp() as i64,
                 seed,
@@ -98,13 +134,16 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
             silver_coins,
             gold_coins,
         } => {
+            let current_transaction_id = vara_man.lazy_get_transaction_id(transaction_id);
             if let Some(game) = vara_man.games.get_mut(game_id as usize) {
                 let player_address = msg::source();
 
+                // Check that game is not paused
                 if vara_man.status == Status::Paused {
                     return VaraManEvent::Error("Incorrect whole game status.".to_owned());
                 }
 
+                // Check that player is registered
                 let Some(player) = vara_man.players.get_mut(&player_address) else {
                     return VaraManEvent::Error("Player must be registered to claim.".to_owned());
                 };
@@ -126,29 +165,14 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
                     return VaraManEvent::Error("Rewards are already claimed.".to_owned());
                 }
 
+                // Check passed coins range
+                if silver_coins > game.silver_coins || gold_coins > game.gold_coins {
+                    return VaraManEvent::Error("Coin(s) amount is gt than allowed.".to_owned());
+                }
+
                 let reward_scale_bps = vara_man.config.get_reward_scale_bps(game.level);
 
-                let gold_coins = gold_coins
-                    .checked_add(
-                        gold_coins
-                            .checked_mul(reward_scale_bps.into())
-                            .expect("Math overflow!")
-                            .checked_div(BPS_SCALE.into())
-                            .expect("Math overflow!"),
-                    )
-                    .expect("Math overflow!");
-
-                let silver_coins = silver_coins
-                    .checked_add(
-                        silver_coins
-                            .checked_mul(reward_scale_bps.into())
-                            .expect("Math overflow!")
-                            .checked_div(BPS_SCALE.into())
-                            .expect("Math overflow!"),
-                    )
-                    .expect("Math overflow!");
-
-                let _tokens_amount = vara_man
+                let base_tokens_amount = vara_man
                     .config
                     .tokens_per_gold_coin
                     .checked_mul(gold_coins)
@@ -162,7 +186,31 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
                     )
                     .expect("Math overflow!");
 
-                // TODO: Transfer tokens
+                let tokens_amount = base_tokens_amount
+                    .checked_add(
+                        base_tokens_amount
+                            .checked_mul(reward_scale_bps.into())
+                            .expect("Math overflow!")
+                            .checked_div(BPS_SCALE.into())
+                            .expect("Math overflow!"),
+                    )
+                    .expect("Math overflow!");
+
+                if ft_messages::transfer_tokens(
+                    current_transaction_id,
+                    &vara_man.config.reward_token_id,
+                    &exec::program_id(),
+                    &player_address,
+                    tokens_amount.into(),
+                )
+                .await
+                .is_err()
+                {
+                    vara_man.transactions.remove(&current_transaction_id);
+                    return VaraManEvent::Error(format!(
+                        "Transaction failed: {current_transaction_id}"
+                    ));
+                };
 
                 player.claimed_gold_coins += player
                     .claimed_gold_coins
@@ -175,6 +223,7 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
 
                 game.is_claimed = true;
 
+                vara_man.transactions.remove(&current_transaction_id);
                 VaraManEvent::RewardClaimed {
                     player_address,
                     game_id,
@@ -182,6 +231,7 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
                     gold_coins,
                 }
             } else {
+                vara_man.transactions.remove(&current_transaction_id);
                 VaraManEvent::Error("Invalid game id.".to_owned())
             }
         }
@@ -196,6 +246,8 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
         VaraManAction::ChangeConfig(config) => {
             if msg::source() != vara_man.config.operator {
                 VaraManEvent::Error("Only operator can change whole game config.".to_owned())
+            } else if !config.is_valid() {
+                VaraManEvent::Error("Provided config is invalid.".to_owned())
             } else {
                 vara_man.config = config;
                 VaraManEvent::ConfigChanged(config)
@@ -207,6 +259,8 @@ async fn process_handle(action: VaraManAction, vara_man: &mut VaraMan) -> VaraMa
 #[no_mangle]
 extern "C" fn init() {
     let init: VaraManInit = msg::load().expect("Unexpected invalid init payload.");
+
+    assert!(init.config.is_valid());
 
     unsafe {
         VARA_MAN = Some(VaraMan {