+++ contracts/vara-man/game/src/lib.rs
@@ -2,6 +2,7 @@
 
 #[cfg(not(feature = "binary-vendor"))]
 mod contract;
+pub mod ft_messages;
 
 /// See `Cargo.toml` for the description of the "binary-vendor" feature.
 #[cfg(feature = "binary-vendor")]