+++ contracts/vara-man/game/tests/utils/mod.rs
@@ -5,4 +5,6 @@ pub use ft::*;
 pub use vara_man::*;
 
 pub const ADMIN: u64 = 100;
+pub const VARA_MAN_ID: u64 = 200;
+pub const FT_ID: u64 = 201;
 pub const PLAYERS: [u64; 3] = [101, 102, 103];