+++ contracts/vara-man/game/tests/utils/vara_man.rs
@@ -1,4 +1,4 @@
-use super::ADMIN;
+use super::{ADMIN, VARA_MAN_ID};
 use gstd::prelude::*;
 use gtest::{Program, System};
 use vara_man_io::{
@@ -8,6 +8,7 @@ use vara_man_io::{
 
 pub trait VaraMan {
     fn vara_man(system: &System) -> Program;
+    fn vara_man_with_config(system: &System, config: Config) -> Program;
     fn register_player(&self, from: u64, name: &str, error: bool);
     fn start_game(&self, from: u64, level: Level, seed: GameSeed, error: bool);
     fn claim_reward(
@@ -26,18 +27,21 @@ pub trait VaraMan {
 
 impl VaraMan for Program<'_> {
     fn vara_man(system: &System) -> Program {
-        let vara_man = Program::current(system);
-        assert!(!vara_man
-            .send(
-                ADMIN,
-                VaraManInit {
-                    config: Config {
-                        operator: ADMIN.into(),
-                        ..Default::default()
-                    }
-                }
-            )
-            .main_failed());
+        Self::vara_man_with_config(
+            system,
+            Config {
+                operator: ADMIN.into(),
+                reward_token_id: ADMIN.into(),
+                gold_coins: 5,
+                silver_coins: 20,
+                ..Default::default()
+            },
+        )
+    }
+
+    fn vara_man_with_config(system: &System, config: Config) -> Program {
+        let vara_man = Program::current_with_id(system, VARA_MAN_ID);
+        assert!(!vara_man.send(ADMIN, VaraManInit { config }).main_failed());
 
         vara_man
     }