+++ contracts/vara-man/game/tests/utils_gclient/mod.rs
@@ -1,0 +1,5 @@
+#![allow(unused)]
+
+pub mod common;
+pub mod ft;
+pub mod vara_man;