+++ contracts/vara-man/CHANGELOG.md
@@ -6,9 +6,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2023-05-15
+### Changed
+- Update dependencies.
+### Added
+- Tokens transfer in claim action.
+- Claim positive and negative tests.
+- Gclient tests.
+
 ## [0.1.0] - 2023-05-01
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/gear-dapps/vara-man/compare/0.1.0...HEAD
-[0.1.0]: https://github.com/gear-dapps/vara-man/compare/6d3ba98...0.1.0
+[Unreleased]: https://github.com/gear-dapps/vara-man/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/gear-dapps/vara-man/compare/0.1.0...0.1.1
+[0.1.0]: https://github.com/gear-dapps/vara-man/compare/1b39920...0.1.0